+++ services/madoc-ts/src/frontend/site/pages/tasks/actions/reject-submission.tsx
@@ -1,3 +1,4 @@
+import { TextField } from '@capture-models/editor/lib/input-types/TextField/TextField';
 import React, { useCallback, useState } from 'react';
 import { useApi } from '../../../../shared/hooks/use-api';
 import { Revisions } from '@capture-models/editor';
@@ -13,6 +14,7 @@ import { DeleteForeverIcon } from '../../../../shared/icons/DeleteForeverIcon';
 export const RejectSubmission: React.FC<{ onReject: () => void; userTaskId: string }> = ({ onReject, userTaskId }) => {
   const api = useApi();
   const [isLoading, setIsLoading] = useState(false);
+  const [reasonMessage, setReasonMessage] = useState('');
   const { currentRevision } = Revisions.useStoreState(state => {
     return {
       currentRevision: state.currentRevision,
@@ -20,21 +22,25 @@ export const RejectSubmission: React.FC<{ onReject: () => void; userTaskId: stri
   });
   const deselectRevision = Revisions.useStoreActions(a => a.deselectRevision);
 
-  const rejectApiCall = useCallback(() => {
-    if (currentRevision) {
-      setIsLoading(true);
-      api
-        .reviewRejectSubmission({
-          revisionRequest: currentRevision,
-          userTaskId,
-        })
-        .then(() => {
-          deselectRevision({ revisionId: currentRevision.revision.id });
-          setIsLoading(false);
-          onReject();
-        });
-    }
-  }, [api, currentRevision, deselectRevision, onReject, userTaskId]);
+  const rejectApiCall = useCallback(
+    (message: string) => {
+      if (currentRevision) {
+        setIsLoading(true);
+        api
+          .reviewRejectSubmission({
+            revisionRequest: currentRevision,
+            userTaskId,
+            message,
+          })
+          .then(() => {
+            deselectRevision({ revisionId: currentRevision.revision.id });
+            setIsLoading(false);
+            onReject();
+          });
+      }
+    },
+    [api, currentRevision, deselectRevision, onReject, userTaskId]
+  );
 
   if (!currentRevision || currentRevision.revision.approved) {
     return null;
@@ -54,6 +60,15 @@ export const RejectSubmission: React.FC<{ onReject: () => void; userTaskId: stri
             <li>The user will be notified that the revision has been rejected</li>
             <li>You will no longer be able to see the content in the revision</li>
           </ul>
+          <label htmlFor="message">Write a message to the contributor</label>
+          <TextField
+            id="message"
+            type="text-field"
+            value={reasonMessage}
+            label="Write message to the contributor"
+            updateValue={setReasonMessage}
+            multiline={true}
+          />
         </div>
       )}
       renderFooter={({ close }: any) => (
@@ -61,7 +76,7 @@ export const RejectSubmission: React.FC<{ onReject: () => void; userTaskId: stri
           style={{ marginLeft: 'auto' }}
           onClick={() => {
             close();
-            rejectApiCall();
+            rejectApiCall(reasonMessage);
           }}
         >
           Reject changes