+++ services/madoc-ts/src/routes/iiif/collections/get-collection.ts
@@ -49,7 +49,7 @@ export const getCollection: RouteMiddleware<{ id: number }> = async context => {
   const returnCollections = [];
   const collection = table.collections[`${collectionId}`] || {
     id: collectionId,
-    label: { '@none': ['Untitled Collection'] },
+    label: { none: ['Untitled Collection'] },
   };
   const manifestIds = table.collection_to_manifest[`${collectionId}`] || [];
   collection.items = manifestIds.map((id: number) => table.manifests[id]);