+++ services/madoc-ts/src/routes/iiif/manifests/get-manifest.ts
@@ -46,7 +46,7 @@ export const getManifest: RouteMiddleware<{ id: string }> = async context => {
 
   const manifest = table.manifests[`${manifestId}`] || {
     id: manifestId,
-    label: { '@none': ['Untitled manifest'] },
+    label: { none: ['Untitled manifest'] },
   };
   const canvasIds = table.manifest_to_canvas[`${manifestId}`] || [];
   manifest.items = canvasIds.map((id: number) => table.canvases[id]);